Replace the get_annotation() function and co with more powerful inference system based on astroid's.
Much more could be said in this commit message, though I won't take this time and one will have to review the diff.

Let's note that it's not properly working yet, the augmented assigments inference code doesn't seem to work at the moment. And we're also missing the unary ops inference.